FE: move /hooks to /app
